+++ XrmPluginCore/Plugins/PluginStepConfig.cs
@@ -14,7 +14,7 @@ namespace XrmPluginCore.Plugins
 
         public string Name { get; internal set; }
 
-        public EventOperation EventOperation { get; internal set; }
+        public string EventOperation { get; internal set; }
 
         public ExecutionStage ExecutionStage { get; internal set; }